# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  handlers.AcknowledgeAlertEventRequest:
    properties:
      eventID:
        type: string
      note:
        type: string
    required:
    - eventID
    type: object
...
  handlers.CreateAlertRuleRequest:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      instance_id:
        type: string
      name:
        type: string
      notification_channels:
        items:
          type: string
        type: array
      notification_config:
        additionalProperties: true
        type: object
      rule_config:
        additionalProperties: true
        type: object
      rule_type:
        enum:
        - threshold
        - anomaly
        - rate
        type: string
      severity:
        enum:
        - critical
        - warning
        - info
        type: string
    required:
    - name
    - rule_config
    - rule_type
    - severity
    type: object
...
  handlers.ErrorResponse:
    properties:
      details: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/handlers.Pagination'
      success:
        type: boolean
    type: object
  handlers.Pagination:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
...
  handlers.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
  handlers.ToggleAlertRuleRequest:
    properties:
      enabled:
        type: boolean
      ruleID:
        type: string
    required:
    - enabled
    - ruleID
    type: object
  handlers.UpdateAlertRuleRequest:
    properties:
      description:
        type: string
      enabled:
        type: boolean
      name:
        type: string
      notification_channels:
        items:
          type: string
        type: array
      notification_config:
        additionalProperties: true
        type: object
      rule_config:
        additionalProperties: true
        type: object
      ruleID:
        type: string
      severity:
        enum:
        - critical
        - warning
        - info
        type: string
    required:
    - ruleID
    type: object
...
host: localhost:12306
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/ysicing/tiga
  description: |
    Multi-instance DevOps management platform with support for MinIO, MySQL, PostgreSQL, Redis, Docker, and Caddy.

    Features:
    - Instance health monitoring and metrics collection
    - Alert management with rules and events
    - Comprehensive audit logging
    - MinIO object storage management
    - Database management with SQL editor
    - Docker container operations
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Tiga DevOps Platform API
  version: "1.0"
paths:
  /api/v1/alerts/events:
    get:
      description: Get a paginated list of alert events
      parameters:
      - description: Filter by alert ID (UUID)
        in: query
        name: alert_id
        type: string
      - description: Filter by instance ID (UUID)
        in: query
        name: instance_id
        type: string
      - description: Filter by status (firing, acknowledged, resolved)
        in: query
        name: status
        type: string
      - description: Filter by severity
        in: query
        name: severity
        type: string
      - description: Start time (RFC3339)
        in: query
        name: start_time
        type: string
      - description: End time (RFC3339)
        in: query
        name: end_time
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List alert events
      tags:
      - alerts
  /api/v1/alerts/events/{event_id}:
    get:
      description: Get alert event details by ID
      parameters:
      - description: Event ID (UUID)
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get alert event
      tags:
      - alerts
  /api/v1/alerts/events/{event_id}/acknowledge:
    post:
      consumes:
      - application/json
      description: Acknowledge an alert event
      parameters:
      - description: Event ID (UUID)
        in: path
        name: event_id
        required: true
        type: string
      - description: Acknowledgement request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AcknowledgeAlertEventRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Acknowledge alert event
      tags:
      - alerts
  /api/v1/alerts/events/{event_id}/resolve:
    post:
      description: Resolve an alert event
      parameters:
      - description: Event ID (UUID)
        in: path
        name: event_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve alert event
      tags:
      - alerts
  /api/v1/alerts/events/active:
    get:
      description: Get all currently firing alert events
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get active alert events
      tags:
      - alerts
  /api/v1/alerts/rules:
    get:
      description: Get a paginated list of alert rules
      parameters:
      - description: Filter by instance ID (UUID)
        in: query
        name: instance_id
        type: string
      - description: Filter by enabled status
        in: query
        name: enabled
        type: boolean
      - description: Filter by severity (critical, warning, info)
        in: query
        name: severity
        type: string
      - description: Search in name, description
        in: query
        name: search
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.PaginatedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List alert rules
      tags:
      - alerts
    post:
      consumes:
      - application/json
      description: Create a new alert rule
      parameters:
      - description: Alert rule creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateAlertRuleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create alert rule
      tags:
      - alerts
  /api/v1/alerts/rules/{rule_id}:
    delete:
      description: Delete an alert rule
      parameters:
      - description: Rule ID (UUID)
        in: path
        name: rule_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete alert rule
      tags:
      - alerts
    get:
      description: Get alert rule details by ID
      parameters:
      - description: Rule ID (UUID)
        in: path
        name: rule_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get alert rule
      tags:
      - alerts
    patch:
      consumes:
      - application/json
      description: Update alert rule details
      parameters:
      - description: Rule ID (UUID)
        in: path
        name: rule_id
        required: true
        type: string
      - description: Alert rule update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateAlertRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update alert rule
      tags:
      - alerts
  /api/v1/alerts/rules/{rule_id}/toggle:
    patch:
      consumes:
      - application/json
      description: Enable or disable an alert rule
      parameters:
      - description: Rule ID (UUID)
        in: path
        name: rule_id
        required: true
        type: string
      - description: Toggle request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ToggleAlertRuleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Toggle alert rule
      tags:
      - alerts
  /api/v1/alerts/statistics:
    get:
      description: Get overall alert statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get alert statistics
      tags:
      - alerts
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...